API Documentation

NICE.NICE — Module

NICE (N-species Ice Table) module – provides simultaneous equilibria solvers:

simulate Net-event kinetic Monte Carlo solver

solve Nonlinear equations solver (Newton trust region method)

The ReactionSystem type is provided as an interface for these solvers.

NICE.ReactionSystem — Type

Definition of the simultaneous equilibria reaction system. This type is the input to the solver functions simulate and solve, and is updated in-place.

The reaction system consists of

$M$ Reactions involving $N$ species;

a matrix $C = \left(c_{mn}\right)$ defining the stoichiometry of the system; rows correspond to species and columns correspond to reactions; each element is the stoichoimetric coefficient of the $n$th species in the $m$th reaction, and is positive if the species is a product, and negative if it is a reactant;

the forward and reverse rate constants $k_f$ and $k_r$ for each reaction;

initial concentrations or activities $\mathbf{a} = \left(a_n\right)$ of each species;

and can be written (through some abuse of notation) as

\[\begin{aligned} \sum_{\left. n \middle| c_{1n} < 0 \right.} &\left|c_{1n}\right| I_n &\xrightleftharpoons[k_r^{\left(1\right)}]{k_f^{\left(1\right)}} \sum_{\left. n \middle| c_{1n} > 0 \right.} &\left|c_{1n}\right| I_n \\ \vdots & ~ & ~ & \vdots \\ \sum_{\left. n \middle| c_{mn} < 0 \right.} &\left|c_{mn}\right| I_n &\xrightleftharpoons[k_r^{\left(M\right)}]{k_f^{\left(M\right)}} \sum_{\left. n \middle| c_{Mn} > 0 \right.} &\left|c_{Mn}\right| I_n \end{aligned}\]

NICE.ReactionSystem — Method

ReactionSystem(stoich, concs, rev_rate_consts, fwd_rate_consts)

Instantiate a ReactionSystem instance using the forward and reverse rate constants for each reaction.

NICE.ReactionSystem — Method

ReactionSystem(stoich, concs, K_eqs; φ)

Instantiate a ReactionSystem instance using the equilibrium constants $K_\text{eq}^{\left(m\right)}$ to derive the forward and reverse rate constants for each reaction:

\[\begin{aligned} k_r &= \frac{\phi}{K_\text{eq} - 1} \\ k_f &= \phi - k_r \end{aligned}\]

NICE.simulate — Method

simulate(rxn_system;
         n_iter=Int(1e+8), chunk_iter=Int(1e+4),
         ε=1.0e-4, ε_scale=1.0, ε_concs=0.0,
         tol_t=Inf, tol_ε=0.0, tol_concs=0.0)

Run a Net-Event Kinetic Monte Carlo (NEKMC) simulation to find the equilibrium concentrations of the reaction system.

NICE.solve — Method

solve(rxn_system, K_eqs; maxiters=1000, abstol=1.0e-9, reltol=0.0)

Solve the system deterministically.

This method solves a linear least squares problem to get an initial guess for $\mathbf{\xi}$,

\[{\left(c_{mn}\right)}^T \left(\xi_m^{\left(0\right)}\right) = \left(a_n - a_n^{\left(0\right)}\right)\]

and then optimizes the equation for the simultaneous equilibria,

\[K_\text{eq}^{\left(m\right)} = \prod_{n=1}^N {\left( a_n^{\left(0\right)} + \sum_{m=1}^M c_{mn} \xi_m \right)}^{c_{mn}}\]

using user-provided $K_ ext{eq}$ values and the Newton's method + Trust Region method.

NICE.solve — Method

solve(rxn_system, K_eqs; maxiters=1000, abstol=1.0e-9, reltol=0.0)

Solve the system deterministically.

This method solves a linear least squares problem to get an initial guess for $\mathbf{\xi}$,

\[{\left(c_{mn}\right)}^T \left(\xi_m^{\left(0\right)}\right) = \left(a_n - a_n^{\left(0\right)}\right)\]

and then optimizes the equation for the simultaneous equilibria,

\[K_\text{eq}^{\left(m\right)} = \prod_{n=1}^N {\left( a_n^{\left(0\right)} + \sum_{m=1}^M c_{mn} \xi_m \right)}^{c_{mn}}\]

using Newton's method + Trust Region.

Instead of specifying the $K_\text{eq}$ values directly, they are approximated here using the forward or reverse rate constants and a parameter $\phi$.